fix: remove allow roles in protected routes
